Restore some methods for safety.
Usually the most of `get_theme_css()` is never used because
it relies on the cached css which the automated transferred theme mods should contain.

However, when the cache is invalid or unset, `get_theme_css()` tries to call
`get_extra_css()` and `css_rule()`, then a fatal error will occur due to
the lack of those methods.

So, we still need these functions as fallback even if they are never used.
Robustness better than briefness.