Disable the big image threshold in anticipation of WordPress 5.3 roll out
WordPress 5.3 comes with a new feature capping the max size of uploaded images. This is not needed on WPCOM due to Photon we are using. This does not seem to be needed on Atomic as well.

A possible side-effects of limitting the max size of images are users unable to upload hi-res images to their sites.

This can safely land prior 5.3 rollout.