Make reactivate-user-plugins more reliable (#385)
This command previously relied on deactivated plugins list that was
saved to a transient, and because atomic transients are saved in
memcache, the list was sometimes lost before Happiness could run
reactivate-user-plugins.

This change use an option rather than a transient. Because wp-cron is
disabled on Atomic, we are not able to schedule clean up of the option,
but we try to clean up the option as much as we can through the use of
the wp-cli `plugin activate` and `wpcomsh reactivate-user-plugins`
commands. In the worst case, if the option is left over from
troubleshooting, it is at least set not to autoload.

@westi suggested this command could be rewritten to be simpler. I'm not
sure if this is indeed simpler, but I believe it is more straightforward
and perhaps more resilient.

resolves #372
resolves #382