Gutenberg on WPCOM/AT: Re-activate the FB and IG block variations (#522)
### What

* Register a callback for the `block.registerType` frontend filter that patches the `Facebook` and `Instagram` `core/embed` block variations so they appear in the inserter and can have their relevant URLs parsed again;
* Enqueue a general GB js script for future GB frontend patches.

Corresponding WPCOM change: https://code.a8c.com/D49605.

### Why

This is relevant for **Gutenberg >= 9.0** (see [CfT](https://wpcomhappy.wordpress.com/2020/09/14/gutenberg-v9-0-call-for-testing/#comment-31039)), where the aforementioned blocks [were deprecated](https://github.com/WordPress/gutenberg/pull/24472) because they do not support the oEmbed changes that Facebook is going to implement on Oct. 24th, 2020, and core has currently no plans to support it.

However, we do plan on keeping support for them on WPCOM/AT. Facebook is adding a token authentication mechanism to its new oEmbed implementation, and we are following suit. We're counting on [this](ttps://developers.facebook.com/docs/plugins/oembed-legacy) to result in something that continues to render the embeds on the backend, using an authenticated mechanism, as required by Facebook's new policy.

Our goal is for this go unnoticed by our end-users, as this is only an implementation detail.

**For now, we're patching them so they continue to work as before, as the current FB oEmbed implementation is still active.**