Masterbar: add the plugins menu to non-business sites (#638)
* Masterbar: add the plans menu to non-business sites

Currently, the plans menu item is only displayed on sites with a business or
ecommerce plan. The Calypso plans page can be used to market plan and marketplace
upgrades to users. So let's add the plugins menu item to non-business/ecommerce sites.
On these sites, the plugins menu item will link to the Calypso plugins page.

On business and ecommerce sites, the plugins menu item will behave as normal: on wp-admin,
it will lead to the wp-admin plugins page while on Calypso it will lead to the Calypso plugins
page.