Stop auto-update-related issues over the weekend (#686)
We just fixed Jetpack plugin auto-updates to work for WoA sites,
but it looks like some sites had previous disabled auto-updates for
plugins and found them re-enabled when switching to core WP auto-update
settings. This might be because of how we force auto-updates for
all plugins in wpcomsh. Either way, we need to stop the bleeding
over the weekend and look at this problem next week.